Redoing the file structure to group components by the pages they're used on
